refactor(ace): use `index.ts` file for exporting
